+++ assets/js/directives/index.js
@@ -1,4 +1,5 @@
 define([
     'plugbot/directives/draggable',
+    'plugbot/directives/href',
     'plugbot/directives/window'
 ], _.identity);