+++ assets/js/services/SiteApi.js
@@ -1,11 +1,11 @@
 define(['plugbot/services/module', 'angular'], function (module, angular) {
     'use strict';
 
-    module.factory('SiteApi', ['$interval', '$window', function ($interval, $window) {
+    module.factory('SiteApi', ['$http', '$interval', '$window', function ($http, $interval, $window) {
         var API = $window.API,
-            elemWoot = angular.element('#woot'),
             idWoot,
-            idJoin;
+            idJoin,
+            listMediaChange;
 
         return {
             woot: function () {
@@ -14,9 +14,11 @@ define(['plugbot/services/module', 'angular'], function (module, angular) {
                 this.stopWoot();
 
                 idWoot = $interval(function () {
-                    var userVote = API.getUser().vote;
+                    var elemWoot = angular.element('#woot'),
+                        user = API.getUser(),
+                        dj = API.getDJ();
 
-                    if (0 === userVote) {
+                    if (dj && dj.id !== user.id && 0 === user.vote) {
                         elemWoot.click();
                     } else {
                         that.stopWoot();
@@ -35,6 +37,33 @@ define(['plugbot/services/module', 'angular'], function (module, angular) {
                         that.stopJoin();
                     }
                 }, 1000);
+            },
+            getMedia: function (callback) {
+                var media = API.getMedia(),
+                    ret;
+
+                if (media) {
+                    ret = {};
+
+                    if (1 === media.format) {
+                        ret.format = 'youtube';
+                        ret.url = '//www.youtube.com/watch?v=' + media.cid;
+                        callback(ret);
+                    } else {
+                        ret.format = 'soundcloud';
+                        $http
+                            .get('//api.soundcloud.com/tracks/' + media.cid +
+                            '.json?client_id=YOUR_CLIENT_ID')
+                            .then(function (data) {
+                                ret.url = data.data.permalink_url;
+                            }, function () {
+                                ret.url = null;
+                            })
+                            .then(function () {
+                                callback(ret);
+                            });
+                    }
+                }
             },
             stopWoot: function () {
                 $interval.cancel(idWoot);
@@ -60,9 +89,37 @@ define(['plugbot/services/module', 'angular'], function (module, angular) {
                     this.stopJoin();
                 }
             },
+            onMediaChange: function () {
+                var that = this;
+
+                _.each(listMediaChange, function (callback) {
+                    callback.onAdvance();
+                    that.getMedia(callback.onResolved);
+                });
+            },
+            bindMediaChange: function (id, onAdvance, onResolved) {
+                if (!listMediaChange) {
+                    API.on(API.ADVANCE, this.onMediaChange, this);
+                    listMediaChange = {};
+                }
+
+                listMediaChange[id] = {
+                    onAdvance: onAdvance,
+                    onResolved: onResolved
+                };
+            },
+            unbindMediaChange: function (id) {
+                if (id) {
+                    delete listMediaChange[id];
+                } else {
+                    API.off(API.ADVANCE, this.onMediaChange);
+                    listMediaChange = {};
+                }
+            },
             destroy: function () {
                 this.autoWoot(false);
                 this.autoJoin(false);
+                this.unbindMediaChange();
             }
         };
     }]);